Add contributing guide
This contributing guide document should be referenced from a brief
CONTRIBUTING.md file in the root directory of a gcloud client
library project. It is general, and links to external docs such
as style guide, set up, etc in order to allow each client library
project to supply these specifics.